requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.31
psycopg[binary]==3.2.9
python-dotenv==1.0.1
waitress==3.0.0
pandas==2.2.2
openpyxl
XlsxWriter
